Child Spawning changes
Released Child Processes from Main Process so that even when the app is closed, the servers will continue to run atonoumsly. This simplifies the shutdown process for the app and I believe it will be more useful than one that forces a window to stay up for long periods of time.